+++ src/OpenTelemetry.Instrumentation.GrpcCore/CHANGELOG.md
@@ -1,7 +1,5 @@
 # Changelog
 
-* Update `OpenTelemetry.Api` to `1.7.0`.
-
 ## Unreleased
 
 * Make the context propagation extraction case insensitive.
@@ -10,6 +8,8 @@
   ([#1344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1344))
 * Update minimal supported  version of `Google.Protobuf` to `3.15.0`.
   ([#1456](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1456))
+* Update `OpenTelemetry.Api` to `1.7.0`.
+  ([#1486](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1486))
 
 ## 1.0.0-beta.5