+++ src/OpenTelemetry.Instrumentation.GrpcCore/CHANGELOG.md
@@ -10,6 +10,8 @@
   ([#1456](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1456))
 * Update `OpenTelemetry.Api` to `1.7.0`.
   ([#1486](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1486))
+* `ActivitySource.Version` is set to NuGet package version.
+  ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
 
 ## 1.0.0-beta.5