+++ src/OpenTelemetry.Instrumentation.GrpcCore/CHANGELOG.md
@@ -10,6 +10,8 @@
   ([#1635](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1635))
 * `ActivitySource.Version` is set to NuGet package version.
   ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
+* RecordException option for both client and server interceptors
+  ([#1648](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1648))
 
 ## 1.0.0-beta.5