+++ src/OpenTelemetry.Instrumentation.GrpcCore/CHANGELOG.md
@@ -10,6 +10,8 @@
   ([#1624](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1624))
 * Update `OpenTelemetry.Api` to `1.8.1`.
   ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))
+* `RecordException` option for both client and server interceptors.
+  ([#1648](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1648))
 * Use `Activity.Status` and `Activity.StatusDescription`
   instead of `otel.status_code` and `otel.status_description` tags.
   ([#1799](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1799))